fix(Interfaces): refactor /shared/interfaces and move to /shared/types
